cli bot command global options
