+++ packages/cli/src/bots.ts
@@ -1,20 +1,22 @@
-import { Command } from 'commander';
-import { medplum } from '.';
-import { createBot, deployBot, readBotConfigs, saveBot } from './utils';
 import { MedplumClient } from '@medplum/core';
+import { createMedplumClient } from './util/client';
+import { createMedplumCommand } from './util/command';
+import { createBot, deployBot, readBotConfigs, saveBot } from './utils';
 
-export const bot = new Command('bot');
+export const bot = createMedplumCommand('bot');
 
 // Commands to deprecate
-export const saveBotDeprecate = new Command('save-bot');
-export const deployBotDeprecate = new Command('deploy-bot');
-export const createBotDeprecate = new Command('create-bot');
+export const saveBotDeprecate = createMedplumCommand('save-bot');
+export const deployBotDeprecate = createMedplumCommand('deploy-bot');
+export const createBotDeprecate = createMedplumCommand('create-bot');
 
 bot
   .command('save')
   .description('Saving the bot')
   .argument('<botName>')
-  .action(async (botName) => {
+  .action(async (botName, options) => {
+    const medplum = await createMedplumClient(options);
+
     await botWrapper(medplum, botName);
   });
 
@@ -22,7 +24,9 @@ bot
   .command('deploy')
   .description('Deploy the app to AWS')
   .argument('<botName>')
-  .action(async (botName) => {
+  .action(async (botName, options) => {
+    const medplum = await createMedplumClient(options);
+
     await botWrapper(medplum, botName, true);
   });
 
@@ -30,7 +34,9 @@ bot
   .command('create')
   .arguments('<botName> <projectId> <sourceFile> <distFile>')
   .description('Creating a bot')
-  .action(async (botName, projectId, sourceFile, distFile) => {
+  .action(async (botName, projectId, sourceFile, distFile, options) => {
+    const medplum = await createMedplumClient(options);
+
     await createBot(medplum, [botName, projectId, sourceFile, distFile]);
   });
 
@@ -50,20 +56,26 @@ export async function botWrapper(medplum: MedplumClient, botName: string, deploy
 saveBotDeprecate
   .description('Saves the bot')
   .argument('<botName>')
-  .action(async (botName) => {
+  .action(async (botName, options) => {
+    const medplum = await createMedplumClient(options);
+
     await botWrapper(medplum, botName);
   });
 
 deployBotDeprecate
   .description('Deploy the bot to AWS')
   .argument('<botName>')
-  .action(async (botName) => {
+  .action(async (botName, options) => {
+    const medplum = await createMedplumClient(options);
+
     await botWrapper(medplum, botName, true);
   });
 
 createBotDeprecate
   .arguments('<botName> <projectId> <sourceFile> <distFile>')
   .description('Creates and saves the bot')
-  .action(async (botName, projectId, sourceFile, distFile) => {
+  .action(async (botName, projectId, sourceFile, distFile, options) => {
+    const medplum = await createMedplumClient(options);
+
     await createBot(medplum, [botName, projectId, sourceFile, distFile]);
   });
